refactor add campsite form
