+++ packages/client/src/components/MapView/MapView.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useEffect, useState, useCallback } from "react";
 import { useSelector, useDispatch } from 'react-redux';
 import { MapContainer, TileLayer, Marker, Popup, Tooltip } from "react-leaflet";
 import L from "leaflet";
@@ -6,6 +6,7 @@ import { fetchCampsites, selectCampsites, selectLoading, selectError } from "../
 import type { AppDispatch } from "../../store/store";
 import "./MapView.css";
 import CampsiteMarker from "../CampsiteMarker/CampsiteMarker";
+import AddCampsiteForm from "../AddCampsiteForm/AddCampsiteForm";
 
 const defaultPosition: [number, number] = [39.2508, -106.2925]; // Leadville
 
@@ -26,6 +27,7 @@ const personIcon = new L.DivIcon({
 });
 
 const MapView: React.FC = () => {
+  const [showModal, setShowModal] = useState<boolean>(false);
   const dispatch = useDispatch<AppDispatch>();
   const campsites = useSelector(selectCampsites);
   const loading = useSelector(selectLoading);
@@ -44,6 +46,10 @@ const MapView: React.FC = () => {
       );
     }
   }, []);
+
+  const handleSuccess = useCallback(() => {
+      setShowModal(false);
+    }, []);
 
   const createCampsiteMarker = () => {
     return campsites.map((site) => (
@@ -92,6 +98,30 @@ const MapView: React.FC = () => {
           </Marker>
         )}
       </MapContainer>
+      <button
+        className="plus-button"
+        onClick={() => setShowModal(true)}
+        aria-label="Add Campsite"
+      >
+        +
+      </button>
+      {showModal && (
+        <div className="modal-overlay" onClick={() => setShowModal(false)}>
+          <div
+            className="modal-content"
+            onClick={(e) => e.stopPropagation()}
+          >
+            <button
+              className="close-modal"
+              onClick={() => setShowModal(false)}
+              aria-label="Close"
+            >
+              &times;
+            </button>
+            <AddCampsiteForm onSuccess={handleSuccess} />
+          </div>
+        </div>
+      )}
     </div>
   );
 };
